refactor(config): centralize editing actions
